anode loader now compatible with anode conv trainer

+++ cosmoNODE/anode/models.py
@@ -1,9 +1,6 @@
-# huge thanks to https://github.com/EmilienDupont/augmented-neural-odes
-# dude is sick
-from math import pi
-
 import torch
 import torch.nn as nn
+from math import pi
 from torchdiffeq import odeint, odeint_adjoint
 
 
@@ -69,16 +66,12 @@ class ODEFunc(nn.Module):
         # Forward pass of model corresponds to one function evaluation, so
         # increment counter
         self.nfe += 1
-        # print(x.shape)
-        # print(x)
         if self.time_dependent:
             # Shape (batch_size, 1)
             t_vec = torch.ones(x.shape[0], 1).to(self.device) * t
             # Shape (batch_size, data_dim + 1)
-            t_vec = t_vec.double()
             t_and_x = torch.cat([t_vec, x], 1)
             # Shape (batch_size, hidden_dim)
-
             out = self.fc1(t_and_x)
         else:
             out = self.fc1(x)
@@ -142,18 +135,15 @@ class ODEBlock(nn.Module):
         if self.odefunc.augment_dim > 0:
             if self.is_conv:
                 # Add augmentation
-                # print(x.shape)
                 batch_size, channels, height, width = x.shape
                 aug = torch.zeros(batch_size, self.odefunc.augment_dim,
-                                  height, width, dtype=torch.double).to(self.device)
+                                  height, width).to(self.device)
                 # Shape (batch_size, channels + augment_dim, height, width)
                 x_aug = torch.cat([x, aug], 1)
             else:
                 # Add augmentation
-                aug = torch.zeros(x.shape[0], self.odefunc.augment_dim).to(self.device).double()
+                aug = torch.zeros(x.shape[0], self.odefunc.augment_dim).to(self.device)
                 # Shape (batch_size, data_dim + augment_dim)
-                # print(x.dtype)
-                # print(aug.dtype)
                 x_aug = torch.cat([x, aug], 1)
         else:
             x_aug = x
